+++ src/DetailsView/reports/report-generator.ts
@@ -6,7 +6,7 @@ import { FeatureFlagStoreData } from '../../common/types/store-data/feature-flag
 import { TabStoreData } from '../../common/types/store-data/tab-store-data';
 import { ScanResults } from '../../scanner/iruleresults';
 import { AssessmentReportHtmlGenerator } from './assessment-report-html-generator';
-import { ReportHtmlGenerator } from './report-html-generator-impl';
+import { ReportHtmlGenerator } from './report-html-generator';
 import { ReportNameGenerator } from './report-name-generator';
 
 export class ReportGenerator {
