+++ src/DetailsView/reports/report-html-generator.tsx
