+++ src/tests/unit/tests/DetailsView/reports/report-generator.test.ts
@@ -7,7 +7,7 @@ import { FeatureFlagStoreData } from '../../../../../common/types/store-data/fea
 import { TabStoreData } from '../../../../../common/types/store-data/tab-store-data';
 import { AssessmentReportHtmlGenerator } from '../../../../../DetailsView/reports/assessment-report-html-generator';
 import { ReportGenerator } from '../../../../../DetailsView/reports/report-generator';
-import { ReportHtmlGenerator } from '../../../../../DetailsView/reports/report-html-generator-impl';
+import { ReportHtmlGenerator } from '../../../../../DetailsView/reports/report-html-generator';
 import { ReportNameGenerator } from '../../../../../DetailsView/reports/report-name-generator';
 import { ScanResults } from '../../../../../scanner/iruleresults';
